Implementacion de metodos en AbstractFacade, modificar POM

+++ src/main/java/sv/edu/ues/fmocc/ingenieria/tpi135/guia02/guia02/AbstractFacade.java
@@ -11,6 +11,7 @@ import javax.persistence.EntityManager;
 /**
  *
  * @author ale
+ * @param <T>
  */
 public abstract class AbstractFacade<T> {
 
@@ -22,17 +23,55 @@ public abstract class AbstractFacade<T> {
 
     protected abstract EntityManager getEntityManager();
 
-    public void create(T entity) {
-        getEntityManager().persist(entity);
+    public T create(T entity) {
+        if (entity != null && getEntityManager()!=null) {
+            getEntityManager().persist(entity);
+            return entity;
+        }else{
+            return null;
+        }
     }
 
-    public void edit(T entity) {
-        getEntityManager().merge(entity);
+    public T edit(T entity) {
+        if (entity != null && getEntityManager()!=null) {
+            getEntityManager().merge(entity);
+            return entity;
+        }else{
+            return null;
+        }
     }
 
-    public void remove(T entity) {
-        getEntityManager().remove(getEntityManager().merge(entity));
+    public T remove(T entity) {
+        if (entity!=null && getEntityManager()!=null) {
+            getEntityManager().remove(getEntityManager().merge(entity));
+            return entity;
+        }else{
+            return null;
+        }
     }
+    
+    public boolean crear(T entity){
+        
+        if (entity!= null && getEntityManager()!=null) {
+            return create(entity)==entity;
+        }
+        return false;
+    }
+    
+    public boolean modificar(T entity){
+        if (entity!= null && getEntityManager()!=null) {
+            return edit(entity)==entity;
+        }
+        return false;
+    }
+    
+    public boolean eliminar(T entity){
+        if (entity!= null && getEntityManager()!=null) {
+            return remove(entity)==entity;
+        }
+        return false;
+    }
+    
 
     public T find(Object id) {
         return getEntityManager().find(entityClass, id);
